(#7)controller 명확하게 하고, service딴 로직분리
